+++ RELEASE_NOTES.md
@@ -1,7 +1,9 @@
 ## Release Notes 3.11.1
 
 ## Bugfixes
+
 - **Tables / PUP Packs**: Fixed wrong order of PUP pack detection when an alias was used.
+- **Tables / VPS Updates**: Fixed filtering issues.
 
 ## Release Notes 3.11.0